needs validation

+++ routes/folders.js
@@ -8,9 +8,17 @@ const mongoose = require('mongoose');
 const Folder = require('../models/folder');
 const Note = require('../models/note');
 
+const passport = require('passport');
+
+//protecting the endpoints
+router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
+
+
+
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-  Folder.find()
+  const userId = req.user.id;
+  Folder.find({userId})
     .sort('name')
     .then(results => {
       res.json(results);
@@ -23,6 +31,8 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
+  
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -30,7 +40,7 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Folder.findById(id)
+  Folder.findOne({_id:id,userId})
     .then(result => {
       if (result) {
         res.json(result);
@@ -46,8 +56,8 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { name } = req.body;
-
-  const newFolder = { name };
+  const userId = req.user.id;
+  const newFolder = { name,userId };
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -73,6 +83,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -87,9 +98,9 @@ router.put('/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateFolder = { name };
+  const updateFolder = { name , userId };
 
-  Folder.findByIdAndUpdate(id, updateFolder, { new: true })
+  Folder.findByIdAndUpdate({_id:id,userId}, updateFolder, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -109,14 +120,14 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
-
+  const userId = req.user.id;
   // ON DELETE SET NULL equivalent
-  const folderRemovePromise = Folder.findByIdAndRemove({ _id: id });  // NOTE **underscore** _id
+  const folderRemovePromise = Folder.findByIdAndRemove({ _id: id, userId });  // NOTE **underscore** _id
   // ON DELETE CASCADE equivalent
   // const noteRemovePromise = Note.deleteMany({ folderId: id });
 
   const noteRemovePromise = Note.updateMany(
-    { folderId: id },
+    { folderId: id,userId },
     { '$unset': { 'folderId': '' } }
   );
