+++ routes/notes.js
@@ -7,12 +7,20 @@ const mongoose = require('mongoose');
 
 const Note = require('../models/note');
 
+const passport = require('passport');
+
+//protecting the endpoints
+router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
+
+
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
-
+  const userId = req.user.id;
+  
   let filter = {};
-
+  filter.userId=userId;
+  
   if (searchTerm) {
     // filter.title = { $regex: searchTerm };
     filter.$or = [{ 'title': { $regex: searchTerm } }, { 'content': { $regex: searchTerm } }];
@@ -40,6 +48,7 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -47,7 +56,7 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findOne({ _id: id, userId })
     .populate('tags')
     .then(result => {
       if (result) {
@@ -64,6 +73,7 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { title, content, folderId, tags = [] } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -88,7 +98,7 @@ router.post('/', (req, res, next) => {
     });
   }
 
-  Note.create({ title, content, folderId, tags })
+  Note.create({ title, content, folderId, tags, userId })
     .then(result => {
       res
         .location(`${req.originalUrl}/${result.id}`)
@@ -104,6 +114,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags = [] } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -134,7 +145,7 @@ router.put('/:id', (req, res, next) => {
     });
   }
 
-  Note.findByIdAndUpdate(id, { title, content, folderId, tags }, { new: true })
+  Note.findByIdAndUpdate(id, { title, content, folderId, tags, userId }, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -150,8 +161,9 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
+  const {userId} = req.user.id;
 
-  Note.findByIdAndRemove(id)
+  Note.findByIdAndRemove({_id:id,userId})
     .then(() => {
       res.status(204).end();
     })
