+++ routes/tags.js
@@ -8,9 +8,18 @@ const mongoose = require('mongoose');
 const Tag = require('../models/tag');
 const Note = require('../models/note');
 
+const passport = require('passport');
+
+//protecting the endpoints
+router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
+
+
+
+
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-  Tag.find()
+  const userId = req.user.id;
+  Tag.find({userId})
     .sort('name')
     .then(results => {
       res.json(results);
@@ -23,6 +32,7 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -30,7 +40,7 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Tag.findById(id)
+  Tag.findOne({_id:id,userId})
     .then(result => {
       if (result) {
         res.json(result);
@@ -46,8 +56,9 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { name } = req.body;
-
-  const newTag = { name };
+  const userId = req.user.id;
+  const newTag = { name , userId };
+  
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -73,7 +84,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
-
+  const userId = req.user.id;
   /***** Never trust users - validate input *****/
   if (!name) {
     const err = new Error('Missing `name` in request body');
@@ -88,8 +99,8 @@ router.put('/:id', (req, res, next) => {
   }
 
   const updateTag = { name };
-
-  Tag.findByIdAndUpdate(id, updateTag, { new: true })
+  console.log('hello');
+  Tag.findByIdAndUpdate({_id:id, userId},updateTag, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -109,11 +120,12 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
-  const tagRemovePromise = Tag.findByIdAndRemove(id);
+  const userId = req.user.id;
+  const tagRemovePromise = Tag.findByIdAndRemove({_id:id,userId});
   // const tagRemovePromise = Tag.remove({ _id: id }); // NOTE **underscore** _id
 
   const noteUpdatePromise = Note.updateMany(
-    { 'tags': id, },
+    { 'tags': id,userId },
     { '$pull': { 'tags': id } }
   );
